Add assertions to various Ui commands
There are several different methods in the Ui class that are called by
the Parser. There is a chance that the wrong method is accidentally
called.

To detect mistakes in the event that the parse command is updated, let
us make some assertions about what the string command was to ensure
that the right method is called by the parser.